TensorFlow Iteration 2
The TensorFlow autonomous program is now able to read for 0 or 4 rings and move to the appropriate zones. Placement of a single ring is not currently functioning. Seems to require a change to the robot's phone mount. The current system is to run the code for Target Zone A if detecting 0 and Target Zone C for any other scenario as Target Zone B's code could not be fully implemented.