TensorFlow Ring Count Accuracy Revision
Robot now lowers the front arm to allow the camera to have better vision of the ring(s). The sampling for finding the average ring count has been moved up to 20 from 10. There was also a pause added before sampling for the ring count. In all this has made the program more accurate at telling the correct amount of rings. The single ring is now able to be detected . Next step is to program route for Target Zone B.